fix component to new layout
